Code: - Create simple snapshots manager class for undo/redo. - Use LZAV compression library when possible. - Use standard assertion fail dialogue when possible. - Use translucent effects the right way. - Fix incorrect realloc in VOIDLIST. - Use callbacks instead of switch for map editing functions.
General:
- Added: Russian translation for offline manual. Thanks to Life_My_Agony.
- Added: All manual packages are now containing sample maps and files.
- Added: Undo/Redo feature for map editing which supports up to 128 actions. Press [CTRL]+[Z] to undo action or [CTRL]+[SHIFT]+[Z] to redo. Available in both edit modes and affects almost any action.
- Added: "MapUndoRestoreCamPos" INI option in [General] section. See description in config chapter.
- Max tiles limit has been raised from 9216 to 30592!
- Removed: Sector merge menu for key [J] when pointing on white wall.
- Improved program stability.
- Various minor updates and fixes.

2D Mode:
- Added: Now is possible to draw wall loops that cover another wall loops properly. The covering loop will become red sector automatically when all of it's points inside of parent sector and all of it's loops which is inside will be transfered in it.
- Added: Loop split tool which allows to quickly split wall loop. Works best with circle shaped sectors. See Loop Split chapter to get more info.
- Classic sector drawing:
    - Added: White walls loop autocompletion feature. It allows to quickly connect new white sector to another. Press [SPACE] to insert point on a white wall. Once status bar shows the message about autocompletion, wait a bit to actually see the future loop or just press [SPACE] on same point again to create it immediately. Note that this feature will not work if there is any line intersections.
    - Added: You can hold [SHIFT]+[SPACE] to insert new point on currently hovered wall vertex without grid snap it.
- Added: You can hold [SHIFT]+[INSERT] to insert new point on currently hovered wall vertex without grid snap it.

3D Mode:
- Auto-create xsectors when applying "Light Phase" and "Z-Phase" in "Sector Tricks" menu.
- Added: "Clean matching" feature in "Sector Tricks" menu. It allows to clean x-sectors properties that matching currently hovered x-sector.
- Removed: Raise/lower floor and ceilings with mouse.

Split Mode:
- Fixed: There was some mouse problems in 3D area when mouse-look enabled.

HUD:
- Fixed: Tile box is not showing on resolutions greater than 640x480 and less than 1024x768.
- Fixed: Damage names in edit dialogs was a bit incorrect.
- Added: Set busy value when changing state of object in edit dialog (useful for panning/lighting).

File Browser:
- Fixed: Potential crash when browsing SEQ/QAV files.

Preview Mode:
- Fixed: Incorrect Z-motion sector behavior when crushing sprites.
- Fixed: Incorrect place of FX effect when selecting middle position in Effect Gen type (modern maps only).
- Added: Laser (type 505) preview (modern features).
- Added: Path Sector SPRITE (type 611) preview (modern features).

+++ src_blood/xmpexplo.cpp
@@ -2792,7 +2792,7 @@ static char getThumbnail_SEQ(char* filepath, int nTile, int wh, int hg, int bg)
 	int nTile2;
 	int i;
 
-	if (fileLoadHelper(filepath, (BYTE**)&pSeq) >= sizeof(Seq))
+	if (fileLoadHelper(filepath, (BYTE**)&pSeq) >= (int)sizeof(Seq))
 	{
 		if (memcmp(pSeq->signature, kSEQSig, strlen(kSEQSig)) == 0 && (pSeq->version & 0xff00) == 0x0300)
 		{
@@ -2824,7 +2824,7 @@ static char getThumbnail_QAV(char* filepath, int nTile, int wh, int hg, int bg)
 	char nRetn = 0;
 	int i, j;
 
-	if (fileLoadHelper(filepath, (BYTE**)&pQav) >= sizeof(QAV))
+	if (fileLoadHelper(filepath, (BYTE**)&pQav) >= (int)sizeof(QAV))
 	{
 		if (memcmp(pQav->sign, kQavSig, strlen(kQavSig)) == 0 && (pQav->version & 0xff00) == kQavVersion)
 		{
